fix: edit goal, digit problem, record investment date, record investment alert, no balance in transfer

+++ database/migrations/2026_05_12_141815_create_goals_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('icon')->default('home');
 
-            $table->decimal('target_amount', 15, 2); 
+            $table->unsignedBigInteger('target_amount'); 
 
             $table->date('target_date')->nullable();
             $table->timestamps();
